Move notebook diffing helpers into separate file
