+++ src/vs/workbench/contrib/notebook/browser/diff/inlineDiff/notebookInlineDiff.ts
@@ -10,13 +10,14 @@ import { IInstantiationService } from '../../../../../../platform/instantiation/
 import { NotebookCellTextModel } from '../../../common/model/notebookCellTextModel.js';
 import { NotebookTextModel } from '../../../common/model/notebookTextModel.js';
 import { INotebookEditorWorkerService } from '../../../common/services/notebookWorkerService.js';
-import { CellDiffInfo, computeDiff } from '../notebookDiffViewModel.js';
+import { CellDiffInfo } from '../notebookDiffViewModel.js';
 import { INotebookEditorContribution, INotebookEditor } from '../../notebookBrowser.js';
 import { registerNotebookContribution } from '../../notebookEditorExtensions.js';
 import { NotebookCellDiffDecorator } from './notebookCellDiffDecorator.js';
 import { NotebookDeletedCellDecorator } from './notebookDeletedCellDecorator.js';
 import { NotebookInsertedCellDecorator } from './notebookInsertedCellDecorator.js';
 import { INotebookLoggingService } from '../../../common/notebookLoggingService.js';
+import { computeDiff } from '../../../common/notebookDiff.js';
 
 export class NotebookInlineDiffDecorationContribution extends Disposable implements INotebookEditorContribution {
 	static ID: string = 'workbench.notebook.inlineDiffDecoration';
